# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Medical Task Tracker API
  version: "1.0.0"
  description: |
    API для управления задачами медицинского персонала.
    Поддерживает создание задач, просмотр по ID и периодам, а также отметку выполнения.
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: tasks
    description: Task management endpoints
paths:
  /api/v1/tasks/create:
    post:
      tags: [tasks]
      summary: Create task
      description: Создаёт обычную или периодическую задачу.
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/tasks:
    get:
      tags: [tasks]
      summary: Get tasks for period
      description: Возвращает список задач в интервале от `from` до `to` (включительно).
      operationId: getTasks
      parameters:
        - name: from
          in: query
          required: true
          description: Дата начала периода. Поддерживаются RFC3339 и YYYY-MM-DD.
          schema:
            type: string
            example: "2026-06-01"
        - name: to
          in: query
          required: true
          description: Дата окончания периода. Поддерживаются RFC3339 и YYYY-MM-DD.
          schema:
            type: string
            example: "2026-06-30"
        - name: status
          in: query
          required: false
          description: |
            Фильтр по статусам. Можно передавать несколько значений:
            `status=planned&status=done` или `status=planned,done`.
          schema:
            type: array
            items:
              type: string
              enum: [planned, done, expired]
          style: form
          explode: true
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/tasks/{id}:
    get:
      tags: [tasks]
      summary: Get task by ID
      operationId: getTask
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int32
        - name: status
          in: query
          required: false
          description: |
            Опциональный фильтр по статусу(ам). Если статус задачи не совпадает
            с переданными значениями, возвращается 404.
          schema:
            type: array
            items:
              type: string
              enum: [planned, done, expired]
          style: form
          explode: true
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [tasks]
      summary: Update task by ID
      operationId: updateTask
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [tasks]
      summary: Delete task by ID
      operationId: deleteTask
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int32
      responses:
        "204":
          description: Task deleted
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/tasks/{id}/tags:
    post:
      tags: [tasks]
      summary: Add tags to task
      operationId: addTaskTags
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskTagsRequest'
      responses:
        "200":
          description: Task with updated tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [tasks]
      summary: Remove tags from task
      operationId: removeTaskTags
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskTagsRequest'
      responses:
        "200":
          description: Task with updated tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Cannot remove required tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/tasks/{id}/done:
    post:
      tags: [tasks]
      summary: Mark task done
      description: |
        Помечает задачу выполненной.
        Для периодической задачи необходимо передать `occurrence_date`.
      operationId: markTaskDone
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            format: int32
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkTaskDoneRequest'
      responses:
        "200":
          description: Task status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoneResponse'
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CreateTaskRequest:
      type: object
      required: [title, description, date, is_recurring, tags]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Провести обход"
        description:
          type: string
          minLength: 1
          maxLength: 1000
          example: "Проверить состояние пациентов в отделении"
        date:
          type: string
          format: date-time
          example: "2026-06-30T10:00:00Z"
        is_recurring:
          type: boolean
          example: false
        tags:
          type: array
          items:
            type: string
          example: ["операции", "отчётность"]
        recurring:
          $ref: '#/components/schemas/CreateRecurringTask'

    UpdateTaskRequest:
      type: object
      required: [title, description, date, is_recurring, tags]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Провести обход"
        description:
          type: string
          minLength: 1
          maxLength: 1000
          example: "Проверить состояние пациентов в отделении"
        date:
          type: string
          format: date-time
          example: "2026-06-30T10:00:00Z"
        is_recurring:
          type: boolean
          example: false
        tags:
          type: array
          items:
            type: string
          example: ["операции", "отчётность"]
        recurring:
          $ref: '#/components/schemas/CreateRecurringTask'

    CreateRecurringTask:
      type: object
      required: [recurring_type]
      properties:
        recurring_type:
          type: string
          enum: [weekly, monthly, yearly, biweekly, shift, parity]
          example: weekly
        end_date:
          type: string
          format: date-time
          nullable: true
          description: Опциональная дата завершения серии. Если не передана, серия считается бесконечной.
          example: "2026-12-31T23:59:59Z"
        recurrence_weekly_rule:
          $ref: '#/components/schemas/CreateWeeklyRuleRequest'
        recurrence_monthly_rule:
          $ref: '#/components/schemas/CreateMonthlyRuleRequest'
        recurrence_yearly_rule:
          $ref: '#/components/schemas/CreateYearlyRuleRequest'
        recurrence_biweekly_rule:
          $ref: '#/components/schemas/CreateBiweeklyRuleRequest'
        recurrence_shift_rule:
          $ref: '#/components/schemas/CreateShiftRuleRequest'
        recurrence_parity_rule:
          $ref: '#/components/schemas/CreateParityRuleRequest'

    CreateWeeklyRuleRequest:
      type: object
      properties:
        week_day:
          type: integer
          minimum: 0
          maximum: 6
          description: День недели (0 - воскресенье, 6 - суббота)
          example: 1

    CreateMonthlyRuleRequest:
      type: object
      properties:
        month_day:
          type: integer
          minimum: 1
          maximum: 31
          example: 15

    CreateYearlyRuleRequest:
      type: object
      properties:
        month:
          type: integer
          minimum: 1
          maximum: 12
          example: 6
        day:
          type: integer
          minimum: 1
          maximum: 31
          example: 30

    CreateBiweeklyRuleRequest:
      type: object
      properties:
        is_odd:
          type: boolean
          description: Нечётная неделя (true) или чётная (false)
          example: true
        week_day:
          type: integer
          minimum: 0
          maximum: 6
          example: 2

    CreateShiftRuleRequest:
      type: object
      properties:
        number_of_task_days:
          type: integer
          minimum: 1
          example: 2
        number_of_shift_days:
          type: integer
          minimum: 0
          example: 2

    CreateParityRuleRequest:
      type: object
      properties:
        is_even:
          type: boolean
          description: Чётные (true) или нечётные (false) дни месяца
          example: true

    TaskTagsRequest:
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          minItems: 1
          items:
            type: string
          example: ["звонок", "операции"]

    MarkTaskDoneRequest:
      type: object
      properties:
        occurrence_date:
          type: string
          description: Для периодических задач. Поддерживаются RFC3339 и YYYY-MM-DD.
          example: "2026-07-07"

    DoneResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [done]
          example: done

    Task:
      type: object
      required: [id, title, description, date, is_recurring, status, tags]
      properties:
        id:
          type: integer
          format: int32
          example: 42
        title:
          type: string
          example: "Провести операцию"
        description:
          type: string
          example: "Подготовить и провести операцию пациенту"
        date:
          type: string
          format: date-time
          example: "2026-06-30T10:00:00Z"
        is_recurring:
          type: boolean
          example: true
        status:
          type: string
          enum: [planned, done, expired]
          example: planned
        recurring:
          $ref: '#/components/schemas/RecurringTask'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'

    RecurringTask:
      type: object
      required: [recurring_type]
      properties:
        recurring_type:
          type: string
          enum: [weekly, monthly, yearly, biweekly, shift, parity]
          example: weekly
        end_date:
          type: string
          format: date-time
          nullable: true
          example: "2026-12-31T23:59:59Z"

    Tag:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "операции"

    ValidationError:
      type: object
      required: [validation_errors]
      properties:
        validation_errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/ValidationError'